Simple React tutorial with Net Ninja. Initial commit.
Kinda did a bit 'fore I caught on i should "git it".

React short how-to:
With nodejs installed, just type:
npx create-react-app name-of-your-app-here

Index html is a static file that will be injected with react components. Just leave it alone,
unless you plan some general template edits.

index.js is the main js file that loads the whole app into index.html's "root" div.
App. js is the file that contains top most app component. By editing this page i changed it's contents.

React components are functions that return html-like jsx code. This code is transpiled by Babel to real html and javascript.
Notice how instead of using html's class attribute we use className. that's because we're still in javascript and class is a
reserved keyword here.

If you put something in {} inside jsx component's "html" it'll be treated as javascript code to be executed
That way you can insert dynamic values as shown in code. Be they references to variables or expressions they will work.
Notice that in a link to gw2 website, the href attribute is not quoted when passed in as a js variable value.
That is another difference in jsx from standard template approach. Just the dynamic value, no need to wrap it in quotes.