Added new components and removed component-based css. All css is now in global index.css file
Well..except one inline style in Navbar.js
Notice the use of double curly brackets. One pair is for containing js code.
The other is for creating object literal that contains css properties.
Keep in mind it's still javascript so we're using javascript version of css property names and values,
such as backgroundColor (js) and not background-color (css).