Captain Hook!
Displaying variables and expression value via templates is cool, but there's one tiny issue:
These don't get updated when displayed variable's value changes. It'll in js, but the template will not update
on it's own to reflect the change.

To get React to ..react.. you can use hooks!
Data stored in a hook is observed and changes to it will be reflected in rendered webpage!

Bitchin'. How we do?
First import the hook:

import { useState } from 'react';

Than declare it:

const [age, setAge] = useState(40);

Yeah you go level up here with destructuring to declare variable names.
Age is the value to be displayed. setAge name of function used to change it's value. And in useState is the initial value.
After that it's just:

setAge(10) to change the age and have everything displaying age update.