Displaying lists in react with hooks.
Complex title for simple thing. Hook blogs contains list of blog meta info in an array.
In the template we just map over the array to display their contents.
One key factor is that each displayed item needs a key property to be clearly identifiable to react engine
so it can properly update when changes are made to source hook. Here key={blog.id} is the part that gets this job done.