Reusing a component.
Like in last commit, but this time in practice.
BlogList is now reused in home component for a second list of blogs.
By using filter method of array, we can filter blogs array to only include those
that meet specified cryteria.

Filter function runs each item in array against a logical condition.
If condition is true the item is kept, if false it's discarded.
The resulting new array is returned as result.

Here i made the second list include all blogs but those written by Zef.