Functions as props.
You get that? Neither do I!
Now seriously, function can be passed as props too..well no shocker there, passing function is nothing new.
Bit fat question is why to do it?

Here we got one example. We want to delete blog from a list in a BlogList component.
But this component just displays the data passed as a prop, it does not hold it.

The data's origin is in home component where it's hook is created and initiated.
And it is operations on hooks that cause react to react and re-render components containing the
changed data (provided key attribute was set, it makes component observe the hook for changes).

If you try to delete a blog in BlogList you'll be operating on passed prop data, but not via methods of hook itself.
You may delete the blog, but the html will not be re-rendered to reflect the change.

By defining the blog delete function in Home component where the hook is defined and available we can
pass it (and it's reference to hook via closure) to child component that can properly operate on hook data
and trigger re-rendering mechanism as it should.