useEffect
This hook is fucking badass!
It's simple as fuck - each time a component re-renders itself the function passed to this hook gets executed.
As for when component re-renders itself - it's when it's initially loaded and when it's state changes (that'd be useState hook's job)

Here you can see it when deleting a blog. Deleting one changes the state, things get re-rendered, and function passed to use effect
gets executed, which you can see in dev console.

One warning, which I already figured out before getting it anyway - if you change state in useEffect function you're asking
for an infinite loop! Changing state causes re-render and that causes useEffect to trigger which fires function
that changes state and so on and on and on...