useEffect dependencies.
Sometimes you don't want useEffect to trigger on each and every change of state, only particular ones.
That degree of control is provided by useEffect dependancies.

Usage is simple - you pass an array of states to useEffect() after the function it'll be triggering.
The  array works on default denial principle. Only states in the array can trigger useEffect hook.
So passing empty one means useEffect will only trigger once - on initial render, and then ignore anything else.
Side note - useEffect itself will only react to actual CHANGES of a state. Trying to re-apply same value
to a given state will not trigger useEffect, dependancies or not.

In the example we created new state name and we passed it to dependancies, unlike the blog state.
Now changes to blogs (like deletion) do not cause useEffect to fire it's function.
But changing the name does!