Json server added
Time for more real life approach. Data will be moved to JSON file which will be observed and served
by JSON server to allow for some REST style calls to outside data source.

The data is in data/db.json file.
You install and run the server with command:

npx json-server --watch data/db.json --port 8000

port needed to be specified as 3000 is the default, but that is alredy occupied by react's runtime server for development.
Once this is up you can access data with endpoints like:

http://localhost:8000/blogs/ - get all blogs
http://localhost:8000/blogs/ - add a blog if using POST method.
http://localhost:8000/blogs/id - get a blog with specific ID (using GET method)
http://localhost:8000/blogs/id - delete a blog with specific ID (using DELETE method)