Custom hooks.
Last time we did fetching data from Json server with error handling.
Now this part of code gets turned into a custom "useFetch" hook.

That way the same fetching code does not have to be re-written each time you need to fetch some data via url.
The hook itself is a function, but additionally it's name must start with "use". Otherwise it'll not be seen as a hook.

Obviously to make it into a reusable hook, the url is now a passed parameter, and a return value must be provided.
Here it's an object with pieces of info - data, isPending and error.

The useStates needed for it were also moved into the hook.
These states cannot be top-level - they must reside inside a function (which is the hook in this case).

Over at home.js side of things, you just need to import the hook and destructure it's return object.
Oh, also feed it an url, cause it's a hungry little fucker and needs one.

Lastly cool little trick. When destructuring the return object from useFetch, you can do this:
{ data: blogs, isPending, error } = useFetch(urlHere)

That :blogs is an alias. That way rest of code doesn't need to change blogs variable name to data,
and you can still show that you're pulling data from a fetch request which you expect to be blogs.