Fixing Fetch for unmounted compontent.
There is one possible data fetch failure that slipped by so far.
When data is fetched properly, but user navigated elsewhere in meantime
and the component that requested the data is already destroyed/unmounted.
That will cause an error to be thrown. You can see this when you click on home link,
and quicky swich to another before it loads (which with 1s timeout before loading date coded in is easy)

To fix this you must implement data fetch abort procedure before component unmounts and adjust error handling.

You start by returning a function in useEffect hook. This is called by the hook when
the component is to be removed. Here you place any cleanup code you want done before component goes puff!

Now, to abort a fetch request in progress on component unmount:

1. Create AbortController instance to terminate the request when it's abort() method is called.
Here we do it on line 9, first thing in our customHook definition.

2. Link said AbortController to the fetch call. Otherwise it won't be able to terminate that particular fetch.
You do this by passing a second parameter to fetch() - an object with optional configuration options. One of them is "signal".
Pass the AbortController to it, and it'll gain ability to iterrupt this fetch() request.

3. Call the AbortController's .abort() method in useEffect's return function (so when component is about to unmount) to kill the request.

This almost takes care of the problem. There was still one loophole left.
Aborting still throws an error. And our custom error handling in useEffect by default changed state.
State of object that was unmounted and therefore once again we try to access something that does not exist.

Conditional error handling was addded, so that the fetch request abort error does not try
to touch the state of a no longer existing component.