Url Parameters.
Here we start using url parameters for fetching particular blogs by their id.
For React to consider a part of url a parameter, rather than address, you need to use
a ":" in a route.

<Route path="/blogs/:id">

Up to colon this is a path. The id is a parameter.
To intercept a parameter from a route we use a new hook - useParams();

import { useParams } from "react-router";

...

const { id } = useParams();

That's all you need to create and intercept url params.
The reset of the changes is simply implementing it in this project.
Adding links with blog id in BlogList component, and creating BlogDetails component
to intercept the id from the url and display a message based on that.

Small css rule was added to prevent underlines under everything in blog lists, since each
blog information panel now became a link and was brutally underlined by default.