React form fields
You can link state of a component with form fields.
However this must be a two way binding. If you just set the value of form fields
to state's value, it'll become unchangeable.

You must provide an onChange function that will update the state with form field's value.

const [title,setTitle] = useState('');

<input type="text" value={title} onChange={(e)=>setTitle(e.target.value)}/>

And that would be that all fine and good...
Except i'm a nosy ass and like to complicate things.
Seems the "set" state function is asynchronous.
Adding a line of console.log after it, that outputs "new" title value actually outputs old one.
The change does happen - output it to html element and it shows it just fine.
But console gets the old value despite the log statement being made after the assignement of field value as new state..